Committing changes previously mentioned.
