+++ Fetch-RecipeBrowser/Views/Main/RecipeDetail.swift
@@ -62,27 +62,8 @@ struct RecipeDetail: View {
             Loader(category: "")
                 .isHidden(details != nil)
             
-            VStack {
-                Text("Failed to load recipe details, try again.")
-                    .padding()
-                    .multilineTextAlignment(.center)
-                
-                Button {
-                    model.fetchError = nil
-                    Task {
-                        await model.fetchRecipes()
-                    }
-                } label: {
-                    Text("Okay")
-                        .padding()
-                        .background(RoundedRectangle(cornerRadius: 4).fill(Color.cyan.opacity(0.2)))
-                }.buttonStyle(.plain)
-
-            }
-            .frame(width: screenSize.width * 0.5, height: screenSize.height * 0.5)
-            .background(RoundedRectangle(cornerRadius: 8).fill(LinearGradient(colors: [.black, .gray.opacity(0.9)], startPoint: .topLeading, endPoint: .bottomTrailing)))
-            .shadow(radius: 4)
-            .isHidden(model.fetchError == nil)
+            LoadFailPopup(model: model, isDetails: true)
+                .isHidden(model.fetchError == nil)
         }
         .background(Color.primary.opacity(0.5))
         .onAppear {
