+++ README.md
@@ -12,9 +12,10 @@ Several features for this project:
 8. Along with the above, XCode Cloud integration is enabled with a quick and simple workflow. Any time changes
     are made to the develop branch in GitHub or pull requests against it are created, the testing suite is run.
 9. The Activity Indicator on the Loader that appears while waiting for API calls to complete is just a UIActivityIndicator that is being used in SwiftUI via a UIViewRepresentable.
-10. There is a 5 second timeout interval on the API requests. A basic retry logic popup will prompt the user to try reloading the recipes/recipe details. 
+10. There is a 3 second timeout interval on the API requests. A basic retry logic popup will prompt the user to try reloading the recipes/recipe details. Note that this is for any individual request, not the whole block.
 11. Users can search for recipes within different categories by selecting the category name in the top right and selecting a new one.
-12. Users can search within a category or within favorites of that category, using the search bar, for a recipe name. 
+12. Users can search within a category or within favorites that have been loaded, using the search bar, for a recipe name.
+** Note that favorites that have not been loaded are not currently saved to disk. Therefore, you will only see favorites from categories you have already looked at on the current session. This is only a problem well outside the scope of this challenge, however (when you have favorites from multiple categories).
 
 Happy testing! Please let me know if you have any questions, I will be happy to answer them.
